cpp--08 - ex01 save state

+++ cpp/cpp08/ex01/Span.cpp
@@ -32,7 +32,7 @@ size_t	Span::shortestSpan() const {
 		size_t minSpan = std::numeric_limits<size_t>::max();
 		size_t span = 0;
 		for (auto it = sortedDeduplicated.begin(); std::next(it) != sortedDeduplicated.end(); ++it) {
-			if ((span = *std::next(it) - *it) < minSpan) {
+			if ((span = static_cast<ssize_t>(*std::next(it)) - static_cast<ssize_t>(*it)) < minSpan) {
 				minSpan = span;
 			}
 		}
@@ -45,5 +45,5 @@ size_t	Span::longestSpan() const {
 		throw std::runtime_error("No span can be found! You need at least 2 numbers in vector.");
 	}
 	auto [minIt, maxIt] = std::minmax_element(numbers_.begin(), numbers_.end());
-	return static_cast<unsigned int>(*maxIt - *minIt);//important to cast
+	return (static_cast<ssize_t>(*maxIt) - static_cast<size_t>(*minIt));//important to cast
 }
